Update App Parts sample
Addresses #1950

* Sample taken to 2.0
* Topic lightly touched; content not addressed
