+++ aspnetcore/mvc/advanced/app-parts/sample/AppPartsSample/AppPartsSample.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../DependentLibrary/DependentLibrary.csproj" />
+    <ProjectReference Include="../Plugin/Plugin.csproj" />
+  </ItemGroup>
+
+</Project>
