+++ aspnetcore/mvc/advanced/app-parts/sample/AppPartsSample/Views/Features/Index.cshtml
