+++ aspnetcore/mvc/advanced/app-parts/sample/AppPartsSample/Views/Home/Index.cshtml
