+++ aspnetcore/mvc/advanced/app-parts/sample/AppPartsSample/Views/Shared/_Layout.cshtml
@@ -1,0 +1,20 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>@ViewData["Title"] - AppPartsSample</title>
+    <environment include="Development">
+        <link rel="stylesheet" href="/css/bootstrap-3.3.7.css">
+    </environment>
+    <environment exclude="Development">
+        <link rel="stylesheet" href="https://ajax.aspnetcdn.com/ajax/bootstrap/3.3.7/css/bootstrap.min.css" asp-fallback-href="/css/bootstrap-3.3.7.min.css" asp-fallback-test-class="sr-only" asp-fallback-test-property="position" asp-fallback-test-value="absolute">
+    </environment>
+</head>
+<body>
+    <div class="container body-content">
+        @RenderBody()
+        <hr>
+    </div>
+</body>
+</html>
