+++ aspnetcore/mvc/advanced/app-parts/sample/AppPartsSample/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using AppPartsSample
+@using AppPartsSample.ViewModels
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
