+++ aspnetcore/mvc/advanced/app-parts/sample/AppPartsSample/appsettings.json
