--- aspnetcore/mvc/advanced/app-parts/sample/global.json
@@ -1,3 +1,0 @@
-{
-    "projects": ["."]
-}
