--- aspnetcore/mvc/advanced/app-parts/sample/src/AppPartSample/.bowerrc
@@ -1,3 +1,0 @@
-{
-  "directory": "wwwroot/lib"
-}
