--- aspnetcore/mvc/advanced/app-parts/sample/src/AppPartSample/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using AppPartSample
-@using AppPartSample.ViewModels
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
