--- aspnetcore/mvc/advanced/app-parts/sample/src/DependentLibrary/project.json
@@ -1,14 +1,0 @@
-﻿{
-  "version": "1.0.0-*",
-
-  "dependencies": {
-    "NETStandard.Library": "1.6.0",
-    "Microsoft.AspNetCore.Mvc": "1.0.1"
-  },
-
-  "frameworks": {
-    "netstandard1.6": {
-      "imports": "dnxcore50"
-    }
-  }
-}
